fix: refactor the code
