+++ src/app/components/login/login.scss
@@ -1,8 +1,75 @@
 @import './../../sass/variables.scss';
 
-.login_page {
+.login_modal {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  letter-spacing: 0.6px;
   padding: 40px;
   background-color: white;
   border-radius: 10px;
   width: 300px;
+}
+
+.modal_header {
+  display: flex;
+  align-items: center;
+  gap: 15px;
+}
+
+.modal_form {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.modal_logo {
+  width: 70px;
+  height: 70px;
+}
+
+.modal_name {
+  font-size: 20px;
+  letter-spacing: 1px;
+  font-weight: 600;
+}
+
+.modal_input {
+  width: 100%;
+  height: 25px;
+}
+
+.modal_input-label {
+  font-family: $font-primary;
+  font-size: 15px;
+}
+
+.modal_visibility {
+  font-size: 12px;
+}
+
+.modal_error {
+  display: none;
+  width: 100%;
+  padding: 5px;
+  font-size: 60%;
+  color: white;
+  background-color: #900;
+  border-radius: 5px;
+  box-sizing: border-box;
+}
+
+.modal_error--shown {
+  display: block;
+}
+
+.visibility_wrapper {
+  display: flex;
+  align-items: center;
+}
+
+.modal_hint {
+  font-size: 1.5rem;
+  font-weight: 700;
+  color: $color-primary;
 }
